qemu: Bump timeout for qmp connection
On a loaded CI system, there can be latency spikes due to
contention on I/O for example.

Since the kola tests already have overall timeouts, bump this timeout
much higher to reduce the flake rate.

If we start going past 30s, it seems likely that instead of bumping
this timeout we need to reduce parallelism on the CI system.