docs: Add cosa container age check in cosa alias
It is easy to miss updating the local COSA container and trigger weird
issues during local development.

Add a check to the cosa shell alias looking at the container creation
time and printing out a warning (with a sleep 10) if it is more than a
week old and thus most probably outdated.

This triggers only if no specific container image is set, as this is the
default and most common case where the latest tag might be out of date.
In all other cases, it is assumed that the user knows what it is doing.

Fixes: https://github.com/coreos/coreos-assembler/issues/2434