mantle/ore: openstack: support setting image protection
This will allow us to protect an image from deletion when we upload
it to VexxHost. If an image is marked as "Protected: true" then you
have to change that setting before you can delete it.

This is mostly part of https://github.com/coreos/fedora-coreos-tracker/issues/964
Now that we will be uploading to VexxHost I'd like to turn on image
protection so we can easily delete/gc our non-production stream
images, which we won't mark with protection.