mantle/kola/harness.go: rerun failed tests once
Sometimes we have CI flakes, sometimes a test is incorrectly classified
as non-exclusive. For both these cases it will be beneficial to rerun
failed tests once in their own VMs. If a test fails in the first run but passes
in the next one, the overall result is still a failure.

https://github.com/coreos/coreos-assembler/issues/2442