fastbuild: Better namespace for refs
I did a `cosa build-fast` in my ostreedev/ostree git repo, which
failed because I'd also imported a container into `tmp/repo` which
creates a ref like `ostree/container/blob...`, but this code tried
to create a ref just named `ostree` - and that would require `ostree`
to be both a file and directory.

A general best practice for refs is namespacing; here, let's use
`cosa/fastbuild/`.