Add `deploy-via-container` and `container-imgref` keys for `image.yaml`
This builds on the `image-format: oci` key to switch to using
the new ostree-container code to do the deployment (via rpm-ostree).

Unlike the above though, this is not something we can turn on
by default until we intend to hard cut over to fetching updates
via containers.

The intention is to use this to produce image builds that are
initially provisioned via a container image so that we can more
easily experiment with the "container native" flow.

A side note; I suspect this will indirectly fix
https://github.com/openshift/os/issues/594
because now we're opening a single large file and pulling that
over virtio instead of lots of smaller files.