[rhcos-4.9] build.sh: pin to kernel < 5.14 to unblock s390x builds
In openshift/os#646 it was reported that the s390x downstream builds
were failing to build the 4K metal image. This seems to be a
regression in the 5.14 version of the kernel.

This changes forces the use of the last successful 5.13 kernel. Of
note, the kernel packages have to be excluded from the `distro-sync`
operation or we end up with two versions of the the kernel installed
in the image.

(cherry picked from commit b3373de2a07c9037eb4a88bf0963a01036553b55)