build+metal: Use shared ostree-importing code
The metal case broke with gangplank+aarch64, because it is
running separate pods for the build or something.

In the Jenkins (x86_64) flow I think we bypass this because we have a single
pod which already has the cached commit in `tmp/repo`.

Anyways, dedup all this code using a shared helper.