cosa2stream: adapt to leading version strings > 2 chars
RHCOS version strings have thusfar lead with a 2 digit identifier that
signifies the OCP version. As we roll into 4.10 time frame, the
handling around that leading version identifier is no longer
sufficient.

This changes the code to split the larger version string by the dots
and then manipulates the leading version so that building the stream
name matches.

For example:

49.84.202110071540-0 -> rhcos-4.9
410.84.202110071540-0 -> rhcos-4.10

Closes: #2479