gangplank: allow for specifying extra build meta to copy
There can be cases where we want to copy over a specific build's
metadata and not just the latest build's metadata. For example,
when we encode information about a parent build that might not be
the latest in the build history. This commit adds another parameter
to the job spec and CLI for copying over an extra set of build
metadata to the target.

This is part of a fix for https://github.com/coreos/coreos-assembler/issues/2380