kola: add KOLA_LEAK_ON_FAIL
Sometimes, we hit test flakes that are hard to reproduce manually or
that are cumbersome to set up the same way a test does. Add a new
`KOLA_LEAK_ON_FAIL` env var which will cause kola to:
- enable console auto-login
- add the SSH key pointed by `KOLA_LEAK_ON_FAIL` to the machine
- not deprovision the machine if the test fails

This allows an engineer to to be able to dig deeper, poking at the VM
when the failure happens, either through SSH, or through the serial
console. So then we could expose this in the pipeline as a parameter so
we can do custom runs with the variable enabled (and manually clean up
any leaked VMs after investigating).

This is more powerful than `--ssh-on-test-failure` because it applies to
any failure in general, including provisioning failures, it allows for
debugging via serial console if SSH itself is broken, and it's feasible
to use in a pipeline.