kola/coretest: factor out function to run cmd and timeout
Add a new `RunCmdTimeout` which runs a command but abandons if it's not
done after a certain timeout. Use it in coretest to dedupe `podman` and
`docker` runs. Additionally, the new code has no chance of leaking
goroutines (which... doesn't really matter here since it's native tests
running via kolet invocations, but just on principle).