ibmcloud: remove meta file for PowerVS and set ovf id to 80
Remove the meta file that previously contained some metadata which was used by the IBMCloud UI to identify the os type for ova images. After discussions with the team it looks like they only recongnize certain types of strings like 'rhel', 'centos', etc.. As of now rhcos/fcos (the strings) are not recognized as supported types and changes need to be made on the powervs side for this.

Also change the ovf id to 80 which is recognzed as rhcos/fcos.