kola/harness.go: fix multiply flag with non-exclusive tests
External tests are executed in test VMs via systemd units. When
we use the multiply flag with non-exclusive tests we duplicate
tests and their systemd units in one VM. This creates services
with the same name.

We fix this by generating the non-exclusive wrapper test before
the tests are duplicated. This way the wrapper test is duplicated
instead of the tests within the wrapper. We avoid having multiple
services with the same name in one VM.