cmdlib: strip any setuid/setgid bits in overlay dirs
In the FCOS pipeline, we run with an unprivileged SCC and compose in an
emptyDir. So OpenShift sets the setgid bit so that we can actually write
to it, which ends up getting into the compose itself. We could upgrade
permissions there, but it seems simpler and more robust to consistently
strip out those bits here. We could probably fold this into ostree as
e.g. `--ugo-mode-bits-only` or something.

See: https://github.com/coreos/fedora-coreos-tracker/issues/1003