kola/tests/network, save mac addresses to /var
During the multiple-nics and bond-with-dhcp, the nic
mac adresses are saved inside the vm on /boot.
Since the way boot is mounted varies in different
platforms (PXE, cosa, etc..), moving to save the macs
in the /var volume.
Doing so will remove differences between ci and production
and will contribute to production stability overall.

Signed-off-by: Ram Lavi <ralavi@redhat.com>