kola: allow skipping of individual basic scenarios
In the downstream RHCOS pipeline, sometimes we are testing with
versions of the `kernel` that are signed with a beta key and fail the
SecureBoot QEMU test. By providing a way to skip the SecureBoot scenario
(or any other scenario), we can be more flexible when building + testing
"unique" versions of RHCOS (or FCOS).