kola/tests/rpmostree: refactor install-uninstall test
The downstream RHCOS pipeline recently started failing because the
`bird` package we were using from EPEL was no longer available. This
is the third time we have had to deal with a package on a remote repo
that no longer worked for the `install-uninstall` use case.

This change moves away from using a remote yum repo and instead does
an install with a locally downloaded RPM. This uses the `aht-dummy`
RPM that was used in the `atomic-host-tests` of yesteryear. It has no
external dependencies and should work across all arches.