Drop "buildroot" image
It's not actually a buildroot image. It was meant as a way to cache
dependencies for faster CI and dev time, but it causes too many issues.
For example, a two-iteration lag to get new packages in, confusion when
building the image re. what you'll actually get, and duplication of
steps between the two Dockerfiles.

Let's go back to what we had before #2290 and we can discuss other ways
to optimize CI and the dev flow.