kola/register: Add a mapping from DepDirs to destination folders
Due to the addition of non-exclusive tests, it is possible
that a single test will have multiple dependency dirs. Since
some tests may have the same dependency directory, non-exclusive
test wrapper may contain the same dependency directory that maps
to different destination directories. So we need to use both
dependency directory and test name as a key to guarantee a unique
mapping.