kola: follow symlinks for kola external data dir
Kola tars and gzips up the data directory and extracts it over an SSH
connection.  If the data directory contains relative symlinks to a
location out of the data directory, it will be a dangling symlink once
it is copied into the test machine. Adding the `-h` flag will follow the
symlinks and adds the actual files. This is useful for things such as
shared libraries between external tests.