mantle/kola: add platform specific timeouts
In some cases, there is significant discrepency between provisioning times
on different platforms (ie. QEMU vs AWS). This forces us to cater timeouts
to the lowest common denominator between the platforms. To circumvent this
issue we will move the timeout so that it is started *after* the provisioning
process.

More specifically, the timeout is started when instances are brought to a
"running" state (i.e ignition begins provisioning). We ensure that another
"provisioning" timeout is checking for flakes between the "provisioning" and
"running" states of an instance.