Dockerfile: link to Prow job and fix chmod hack
We currently need to hardcode the base Fedora image on the Prow side as
well, so let's add a comment in the `Dockerfile` as a reminder that it
needs to be bumped at the same time.

While we're here, fix the `/go` hack and add a comment to explain more.

See: https://github.com/openshift/release/pull/23744