build: Ship `ci/` directory installed
The Prow flow is really oriented towards building a container
and then testing it.  It's really maximally convenient if that
container ships the test code too.  This is distinct from what
we're doing with Jenkins which makes it easy to just inject code
into a running a container.

coreos-assembler is huge, these scripts are tiny, so let's just ship
them unconditionally.  Prep for testing cosa `rhcos-4.9` via OpenShift
Prow.