kola/platform: compress userdata before passing to aws
Recently non-exclusive tests have grown to a point such that
the associated ignition config has become too large to be passed
to aws via user-data. The base64 encoding of the config is larger
than 16 KB. While we look for a more permanent solution to this
issue, compressing the config will decrease how often we run into
it. It will also unblock currently broken non-exclusive tests on aws.