kola/external-tests: add support for `minDisk` field
For the Kubernetes node external test, we want to be able to run it in
both QEMU and AWS. But it's a storage heavy test, so the default 10G
image size is not enough. There's `additionalDisks` which works on QEMU,
but not on AWS. But really, we don't need it to be an additional disk.
Having the root disk itself be larger works just as well.

Add a new `minDisk` knob for this and plumb it through QEMU and AWS.