mantle/platform/aliyun: bump default request timeouts
During recent testing of the Aliyun stage of the RHCOS pipeline, we
ran into a timeout error:

`Copying image to region "ap-northeast-1": copying image: [SDK.TimeoutError] The request timed out 4 times(4 for retry), perhaps we should have the threshold raised a little? Read timeout. Please set a valid ReadTimeout.`

This bumps the ReadTimeout to 30s and while we are here, bump the
ReadTimeout to 15s.  (Both are a 3x increase from the API defaults.)