buildextend-live: add features.json to live initrd image
When coreos-installer {iso|pxe} customize runs, it would like to know
what features are supported by the OS image it's customizing.  Most of
that information is inside the squashfs, where coreos-installer can't
easily get to it.  Collect that info (currently only one flag) and write
a JSON file inside the initramfs image, where coreos-installer can find it
in both the ISO and PXE images, and where it's not too easy for other
random code to start depending on it.