kola/harness.go: fix failed tests not being rerun
Currently, a test is checked for failure after the runTest function
has finished running. However, it is possible that a test is marked as
failed after the execution of the runTest function is completed. For example,
a test takes too long to clean up resources and a timeout occurs; the test will
be marked as failed after we check whether the test should be rerun.

To fix this problem, let's only check whether a test has failed after we are
certain that the failed or passed value cannot be modified. So, we will check
for failures after the test suite has finished running.