testiso: add --add-nm-keyfile for NM keyfile propagation testing
With `--add-nm-keyfile`, we inject a NM keyfile into the ISO and enable
`--copy-network`. We then add a systemd service to verify that it took
effect.

Essentially, this verifies that with `coreos-installer iso network
embed` and `--copy-network`, it's possible to specify the NM keyfile
*once* and have that apply to: the live initrd, the live rootfs, the
target initrd, and the target rootfs.

I debated whether to make this a scenario instead of a switch. I ended
up making it a switch because it's another "axis" to the install
(similarly to `--qemu-multipath`). And while this currently only take
effect on ISO installs, nothing stops us from also supporting it for
PXE, even though it's less relevant there (hence why I left it as a
follow-up for now).