buildextend-live: add features.json directly to ISO
The ISO and PXE images share an initramfs, which ships features.json
so `coreos-installer pxe customize` can read it.  This was thought to be
sufficient for `coreos-installer iso customize` as well, and it is, but
it requires `iso customize` to decompress the entire initramfs just to
extract that file.

To avoid that, add a second copy directly to the ISO.  This wastes 2 KiB
of space and makes features.json a more visible API, but we already have
other semi-private APIs in the same ISO directory (kargs.json,
miniso.dat).

On my system, with a release build of coreos-installer, this saves 1.25
seconds for a no-op `iso customize` run.