mantle/platform: azure: allow for not passing SSH key through API
Azure requires that either a username/password be set or an SSH key.
Previously we were unconditionally passing an SSH key but we have tests
(like coreos.ignition.ssh.key) that don't want an SSH key to be passed
via the API and set a NoSSHKeyInMetadata flag for that.

Let's start passing a username/password via the API so we can then bring
up instances without having to specify an SSH key. The username/password
we give won't be used because we don't install their agent.