mantle/ore: add visibility command for Aliyun
This introduces a new subcommand for `ore aliyun` called `visibility`
that takes one or more `region:image` pairs and one of the
`--public|--private` flags. The command will configure the visibility
of the image in the Aliyun cloud appropriately.

The API around this is frustrating for a couple of reasons, which led
me to create a new subcommand:
  - you cannot change the sharing permissions of an image with any
    other function in the SDK
  - copied images do not inherit the public sharing permission from
    the original image
  - image IDs are unique to regions, so you must provide the
    appropriate region with the image ID

The existing subcommands infer a single region from the user's
profile/credentials for the API calls that require a region, which
made it difficult to extend those subcommands to support multiple
regions.