tools: update go modules based on `go mod tidy`
Dependabot has been sending alerts for CVEs on some of the modules,
but has been unable to resolve them automatically.

Specifically:

https://github.com/advisories/GHSA-q6gq-997w-f55g
https://github.com/advisories/GHSA-25xm-hr59-7c27

After speaking with folks internally, `go mod tidy` appears to drop
the requirement on the vulnerable version of `ulikunitz/xz` which
hopefully will quiet the alert for those CVEs.