kola/upgrade/basic: drop metrics.promsock check
IIRC, it started initially as a way to sanity-check that Zincati
auto-updates were turned on and that we wouldn't just wait in vain for
it to update.  But for some reason, sometimes `socat` simply fails to
find Zincati's UNIX socket, even though the latter says that it created
it and goes on with staging the update. For more information, see:

https://github.com/coreos/fedora-coreos-tracker/issues/756.

We should dig into this further, but strictly speaking, it's not even
necessary for this test to have the socket available. So let's just nuke
that check. The `runFnAndWaitForRebootIntoVersion` function will still
time out after 2 minutes if the machine hasn't rebooted by then.