tools: remove all the go module cruft
The content in the `tools` directory is used to build a number of
binaries from source, in order for use with the `gangplank` build
process.

Rather than have to manage golang modules to just get binaries, let's
start installing the binaries directly rather than building directly
from source/modules.

The removal of all the go module scaffolding is the first step towards
installing the tools as binaries.