mantle/kola: retry rpm-ostree status for extra info
We have a case where the rpm-ostree status command is failing
for the ostree.hotfix test and we don't know why:
https://github.com/coreos/fedora-coreos-tracker/issues/942

Let's run it once and check the failure. If it fails we will
always return a failure because we want to know, but we will
also run it with some retries to see if it succeeds in a
successive try for some reason or if it continues to fail.