kola: add subtest runner which logs entry and exit
Use the new `JournalLog` function to add a new `RunLogged` function to
`TestCluster` which has the same semantics as `Run`, i.e. starting a
subtest, but additionally logs in the machines' journals on entry and
exit of the subtest.

Use this in `ostree.basic` and `ostree.hotfix`, the latter of which has
been giving us issues. This should make it easier to read the logs.