mantle: add new RetryUntilTimeout() func, use for early SSH commands
Looking at the function signature of util.Retry (A) and the way
we were calling it (B)

    A) func Retry(attempts int, delay time.Duration, f func() error) error
    B) util.Retry(sshRetries, sshTimeout, f)
        - where:
          - sshRetries = 60
          - sshTimeout = 10 * time.Second

it seems as if our intention was to be able to retry the SSH command
up to 60 times having each one of those 60 tries timeout after 10
seconds. In reality what happens is that we run the ssh command,
which eventually times itself out and then delay 10 seconds before
running the ssh command the next time (60 times). In my experience
with [1] the ssh command would take around 35s to time itself out.
This means Retry() could take 60*(35s + 10s)=2700s=45m to actually
stop trying.

I don't think this was the original intention. Let's make this easier
to follow by adding a RetryUntilTimeout() function and use that
instead. Also let's get rid of the global variables as I think they
aren't really needed.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1066#issuecomment-1009978326