mantle/{kola,platform}: include Ignition runtime in test timer
platform.StartMachine() blocks on SSH being available:

StartMachine()
     |-> StartMachineAfterReboot()
         |-> j.Start()
             | -> util.Retry() <- ssh retry

So if we call it as part of our NewMachine() bringup there's
no way to start the test execution timer and have it include
the time it takes to execute Ignition (which has to complete
before SSH will become available).

Let's add a Machine Option (SkipStartMachine) that will tell the
NewMachine code to skip the call to StartMachine() and return as
soon as the platform thinks the machine is running. Then in our
test code we can start the timer and then call StartMachine()
(via the new mach.Start() interface) manually to set up journal
forwarding and perform sanity checks.

This should help with issues like one we had recently [1] where our
timeouts weren't properly timing out. This was because the provider
timeout (implemented via WaitUntilReady()), had already moved out of
that code but the test timeout (via ExecTimer) hadn't started yet
because everything was blocking under StartMachine().

NOTE: This commit doesn't solve the problem for tests that bring up
      machines themselves by calling NewMachine() but should solve it
      for the large majority of tests.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1066#issuecomment-1009978326