mantle/aliyun: allow CopyImage to wait for image readiness
In my experience, trying to operate on an image immediately after it
had been copied to a region can be unreliable. The `CopyImage()` API
call is "fire and forget", with no native way to monitor the status of
the request. (Unlike `ImportImage()`, which returns a TaskID that
*can* be monitored).

This change introduces a new function that checks the status of an
image until it becomes available or a timeout (10min). This is made
available to the `ore aliyun copy-image` command via a new
`--wait-for-ready` flag that causes the command to wait until the
image can be operated on.