qemuexec: automatically resize the terminal on serial
One really annoying thing when using the serial console is that you have
to manually set (and update if you resize your terminal) the terminal
size.

I found
https://wiki.archlinux.org/title/Working_with_the_serial_console#Resizing_a_terminal
which has a function to resize the terminal which seems to work well. It
doesn't fix everything (e.g. you still get commands overwriting each
other if it gets too long), but helps a lot with for example command
output.

Create a new `autoresize` Ignition fragment which shoves a Bash-adapted
version of the function in a `profile.d` dropin and gets called on every
command prompt.  Automatically inject the fragment when using serial
console devshell.