tests/ostree/unlock: use G_MESSAGES_DEBUG=rpm-ostreed
`all` means debug logging for the whole stack, and ostree can be really
noisy when traversing commits. This makes archived test logs fat enough
that we're approaching the PVC cap in prod.

Of course, eventually we'll remove this once we fix that flake but
short-term let's at least reduce logging to what we need.