kola/devshell: clip status text to terminal width
During SSH bringup, we print the serial console output as it happens.
This output is currently clipped at a hardcoded 100 characters. So on
smaller screens/tmux panes, we'll spill over new lines easily.

Instead, query the terminal size and use that. While we're use, tweak
things in `displayStatusMsg` so that the clipping happens last so we
account for e.g. the status bit and stripped control characters.