cosalib/ova: template VMware HW and OS versions from image.yaml
FCOS and RHCOS want to set different VMware hardware versions and OS
type IDs.  Read both from image.yaml, defaulting to the values we
currently use.

OVF images created by VMware Workstation 16.2.3 respectively record
Fedora and RHEL 8 guests as:

    <OperatingSystemSection ovf:id="100" vmw:osType="fedora64Guest">
    <OperatingSystemSection ovf:id="80" ovf:version="8" vmw:osType="rhel8_64Guest">

The ovf:id values are defined to come from CIM_OperatingSystem.OsType,
which defines 80 as "RedHat Enterprise Linux 64-Bit" and 100 as
"Linux 2.6.x 64-Bit".  (2.6.x seems to be the last explicit kernel version
listed in the spec.)  ovf:version appears to be the major version of the
OS.  VMware's ovftool doesn't seem to care about either of those fields,
only the vmw:osType, so we hardcode the ovf:id to 100 and call it a day.