create_disk: Pull ostree-container over 9p, instead of pull-local
This should fix https://github.com/openshift/os/issues/594

Basically I think 9p has a bug where under memory pressure, having
it *free* an inode requires allocation, which can fail.

This works around that bug by pulling the ostree-container archive
instead, which is a single big file.

Note that this code path does *not* change the semantics at all
for the generated disk.  The information about the pulled container
is discarded and lost.  Actually making use of the container bits
natively is the `deploy-via-container` image option which is
still experimental, but will be used when we progress the
ostree native containers work.