mantle/network/ssh: use ExtendedAgent for SSHAgent
The `Agent` interface doesn't support `SignWithFlags`. So when trying to
use `ssh` with the mantle agent to SSH into a VM, the signing request
during handshake has its flags stripped:

https://cs.opensource.google/go/x/crypto/+/3147a52a:ssh/agent/server.go;l=133-137

This causes the wrong algorithm to be used, which leads to:

    agent key RSA SHA256:<hash> returned incorrect signature type

The reason kola itself doesn't hit this is because it doesn't go through
the socket.

The underlying `keyring` object does implement `ExtendedAgent` though,
which does support `SignWithFlags`, so just cast it.

Fixes: dcd432d2a ("mantle: switch back to RSA SSH keys")