build: Support `builds/` in workdir being a symlink outside workdir
This is the case in the RHCOS pipeline today, and we can support
it in the same way we support `src/config` being a symlink to
outside the workdir.

Note: This only works for absolute symlinks today.

Side note: At some point we should use virtiofs which AIUI will
just fix this by supporting symlinks natively.

Closes: https://github.com/openshift/os/issues/754