Makefile: check that schema copies are synced
On a number of occasions, we've forgotten to synchronize the various
copies of v1.json and the files generated from them.  We should reduce the
number of copies we carry, but for now, have CI check that all copies are
in sync.  Add source hashes to the generated files so we can check that
the latter are up-to-date without regenerating them.