cmdlib: Drop dead "is fedora" or "is RHEL" detection code
I was trying to do stuff related to C9S and this code was just
in my way.  Detecting specific distributions this way is really
an anti-pattern and we should stop it.  Most of this should
be replaced with feature detection, but in *this* case the
code is just a dead relic of a previous attempt to do a RHEL8-based
coreos-assembler.

(I personally do think a C9S-based coreos-assembler could make
 sense, but that's a future topic)