tests/rhcos/upgrade: drop use of sudo for unpriv pods
When running the `rhcos.upgrade.from-ocp-rhcos` test in an
unprivileged pod in the Prow clusters, the test fails during the
unpacking of the `oc` binary because it is trying to use `sudo` to
write to `/usr/bin`.

But since https://github.com/coreos/coreos-assembler/pull/2777 we have
the `oc` binary installed in `cosa`, so just drop the code doing the
install of the binary.

(cherry picked from commit 37f6c2a5e08f46875a986c6185daeb4c2169d5c4)