cmdlib: set permissions on /usr/share/buildinfo
When building RHCOS and generating the `content_manifest.json` file,
the permissions on the file included the `g+w` bit, which causes the
file to be flagged by the `ext.config.shared.files.file-directory-permissions`
test. Add a step in the creation of the ostree layer to set the permissions
on the file before the ostree layer is committed.

Hat tip to @jlebon for the help!