push-container: add option to extend image name
In the RHCOS pipeline, we want to be able to push the legacy
oscontainer and the new base-image container using the same image
reference (i.e. `quay.io/openshift-release-dev/ocp-v4.0-art-dev:411.86.202205021327-0-x86_64`).

When the base-image container is pushed to the same `image:tag`
reference, it overwrites the existing digest (as expected). But this
ends up breaking some of the ART automation around the
`machine-os-content` promotion, because the `sha256` digest in
`meta['oscontainer']['digest']` is still pointing to the digest from
the oscontainer push.

To workaround this for now, let's just provide a flag to append the
`-base-image` string to the image name of the base-image container that
is getting pushed.

i.e. `quay.io/openshift-release-dev/ocp-v4.0-art-dev:411.86.202205021327-0-x86_64-base-image`