kola/tests: drop DO exclusion on Ignition network tests
Container Linux didn't have full networking in the initramfs on
DigitalOcean, so had to disable some tests.  We don't currently test or
ship on DO, and if we ever do, hopefully we'll fix that problem.
Optimistically remove the exclusion so we don't forget about it if we
start testing there.