mantle/platform/api/gcloud: drop interactive OAuth flow
The OAuth client ID corresponds to an app registration that's been dead
for a while now.  We can't usefully create a new registration because the
out-of-band credential flow has been disabled for new apps:

    https://developers.googleblog.com/2022/02/making-oauth-flows-safer.html#disallowed-oob

The recommended alternative is to have OAuth redirect to a web server
listening on localhost, but we usually run in a container, so that would
be awkward.  In practice, we just use --gce-json-key these days, so it's
not worth trying to fix this functionality.