lib: Handle explicit `skip_compression` option in VARIANTS
Add a `skip_compression` property on `QemuVariantImage`  based on
the Set `skip_compression` value of the variant in `VARIANTS` dict.
When `skip_compression` is set to True (default is False), the
`mutate_image` method adds `skip-compression:True` to the `meta_patch`.

This commit also sets `skip_compression` value to True for Nutanix variant.