tests: fix /var race in multipath.partition part 2
This fixes the same race condition in a previous PR #2457 [0]. The
difference is that in PR #2758 [1], `mpath-var-lib-containers.service`
uses mkdir to create `/var/lib/containers` and in some cases it will race
against `ostree-remount.service` during the short window `/var` is
read-only. Have `mpath-var-lib-containers.service` run after
`ostree-remount.service` to resolve this.

[0] https://github.com/coreos/coreos-assembler/pull/2457
[1] https://github.com/coreos/coreos-assembler/pull/2758