create_disk: fix literal `$ignition_firstboot` karg on s390x
While reviewing some related code, I realized that `$ignition_firstboot`
was literally being embedded into the kernel command-line on s390x. We
don't use GRUB there, so it will never get evaluated. Having it there
doesn't break anything but it's useless and can cause confusion, so
remove it.

We already handle the `ignition.firstboot` karg in all the places we
need to for s390x.