build: Use `cache=unsafe` for `cache.qcow2` and transient builds
Followup to the addition of `--transient` for CI flows.  If
we had https://github.com/coreos/rpm-ostree/pull/3719
this would mostly be unnecessary, but today in CI and prod builds,
ostree is invoking `fsync()` inside the supermin VM, and then
qemu in the container is going to `fsync()` all changes down to
the host system (overlay)fs.

For transient flows, switch to `cache=unsafe` so we stop doing
that which should greatly help speed.