azure/instance.go: make CreateOrUpdate timeout aware
In a recent test run in FCOS CI, we noticed CreateOrUpdate hung
and took 30 minutes to terminate. Let's pass the supported cancel
channel to CreateOrUpdate to make it timeout aware.

Note: The version of Azure Go SDK we use (v8.1.0-beta) uses the
deprecated field http.Request.Cancel to provide this functionality.