cmd-push-container: fix --name-suffix semantics
I misunderstood the original intent of this switch while reworking it.
The idea is to add a suffix to the image tag, not its name.

Fix it, and rename the switch accordingly.

While we're here, add a metavar to `name` to make it clear that an
optional tag can also be provided.

Fixes c462739e1 ("cmd-push-container: replace --base-image-name by
--name-suffix").

Closes: https://github.com/openshift/os/issues/847